+++ frontend/public/changelog.md
@@ -10,6 +10,7 @@
 - Membuat hierarki kata dan menyesuaikan data untuk memfasilitasi itu
 - Menambahkan Google Analytics untuk melacak penggunaan situs
 - Menambahkan batasan pencarian 60 kali/15 menit untuk mencegah penyalahgunaan
+- Menerapkan autentikasi oAuth Google
 
 ## 2026-02-15 Minggu